requirements.txt
Flask==1.0.0
Flask-Limiter==0.9.1
Flask-RESTful==0.3.5
psycopg2-binary==2.7.3.2
flask_restful_swagger==0.19